Moves DiscardTransformEngine to the transform package.
