+++ src/org/jitsi/impl/neomedia/transform/DiscardTransformEngine.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.impl.neomedia;
+package org.jitsi.impl.neomedia.transform;
 
+import org.jitsi.impl.neomedia.*;
 import org.jitsi.impl.neomedia.rtp.*;
-import org.jitsi.impl.neomedia.transform.*;
 
 import javax.media.*;
 import java.util.*;
